# gateway/swagger.yaml
openapi: 3.0.1
info:
  title: Learning Platform API (Gateway)
  version: "1.0.0"
  description: "Gateway that proxies to auth, courses, categories, student services."
servers:
  - url: http://localhost:5000
    description: Local Gateway
paths:
...
  /categories:
    get:
      summary: Get all categories
      tags: [Categories]
      security:
        - bearerAuth: []
      responses:
        "200": { description: list of categories }
    post:
      summary: Create category (admin)
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201": { description: created }
  /students/courses:
    get:
      summary: Get courses (via course service)
      tags: [Student]
      security:
        - bearerAuth: []
      responses:
        "200": { description: list of courses }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...